Finalizing AccountRepository, objects, interface and migration

+++ database/migrations/2020_05_29_124231_create_salt_edge_accounts_table.php
@@ -19,7 +19,7 @@ class CreateSaltEdgeAccountsTable extends Migration
             $table->softDeletes();
             $table->bigInteger('account_id');
             $table->string('account_name');
-            $table->integer('salt_edge_customer_id', false, true);
+            $table->bigInteger('salt_edge_customer_id', false, true);
             $table->foreign('salt_edge_customer_id')->references('id')->on('salt_edge_customers')->onDelete('set null');
             $table->binary('object');
             $table->string('hash');
